Добавляет webpack и разбивает скрипт на модули
