+++ src/components/read-more-button.js
@@ -1,0 +1,3 @@
+export const getReadMoreButtonMarkup = () => `
+  <button class="films-list__show-more">Show more</button>
+`;
